refactor: change backend auth logic and separate some classes
